requirements.txt
apache-airflow==2.9.3
pandas==2.2.2
requests==2.31.0
psycopg2-binary==2.9.9
pyarrow==17.0.0